requirements.txt
adjustText==1.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.5
bitarray==3.7.1
bs4==0.0.2
cbor2==5.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
ckzg==2.1.4
contourpy==1.3.3
crytic-compile==0.3.10
cycler==0.12.1
cytoolz==1.0.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fonttools==4.60.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
hexbytes==1.3.1
idna==3.10
kiwisolver==1.4.9
lizard==1.17.31
matplotlib==3.10.6
multidict==6.6.4
numpy==2.3.2
packaging==25.0
pandas==2.3.2
parsimonious==0.10.0
pathspec==0.12.1
pillow==11.3.0
playwright==1.55.0
prettytable==3.16.0
propcache==0.3.2
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyunormalize==16.0.0
regex==2025.9.18
requests==2.32.5
rlp==4.1.0
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
slither-analyzer==0.11.3
smmap==5.0.2
solc-select==1.1.0
soupsieve==2.8
toolz==1.0.0
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
web3==7.13.0
websockets==15.0.1
yarl==1.20.1
